+++ backend/src/main/java/com/ssafy/togeball/domain/league/entity/Stadium.java
@@ -15,7 +15,7 @@ public class Stadium {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="stadium_id")
-    private byte id;
+    private Integer id;
 
     @OneToMany
     private List<Club> clubs;
